+++ source/nodejs/adaptivecards-visualizer/package.json
@@ -17,7 +17,6 @@
     "version": "npm run build"
   },
   "devDependencies": {
-    "@types/ace": "0.0.32",
     "@types/node": "0.0.2",
     "@types/vkbeautify": "^0.99.2",
     "autoprefixer": "^6.7.7",
@@ -33,7 +32,8 @@
     "webpack": "^3.8.1"
   },
   "dependencies": {
-    "brace": "^0.9.1",
-    "adaptivecards": "^1.0.0-0"
+    "adaptivecards": "^1.0.0-0",
+    "markdown-it": "^8.4.0",
+    "monaco-editor": "^0.10.1"
   }
 }
