+++ source/nodejs/adaptivecards/webpack.config.js
@@ -38,8 +38,5 @@ module.exports = {
                 loader: "json-loader",
             }
         ]
-    },
-    externals: {
-        "markdown-it": { var: 'markdownit' }
     }
 };
